Connectors: Update Finance app connector
