+++ packages/connect-finance/src/types.ts
@@ -1,6 +1,22 @@
-import { SubscriptionHandler } from '@aragon/connect-types'
-import Transaction from './entities/Transaction'
-import TokenBalance from './entities/TokenBalance'
+import { SubscriptionHandler, Address } from '@aragon/connect-types'
+import Transaction from './models/Transaction'
+import TokenBalance from './models/TokenBalance'
+
+export interface TransactionData {
+  id: string
+  token: Address
+  entity: Address
+  isIncoming: boolean
+  amount: string
+  date: string
+  reference: string
+}
+
+export interface TokenBalanceData {
+  id: string
+  token: Address
+  balance: string
+}
 
 export interface IFinanceConnector {
   disconnect(): Promise<void>
